start refactoring by moving validators into own folder & namespace
