+++ Rakefile
@@ -8,21 +8,21 @@ RSpec::Core::RakeTask.new
 task :default => :spec
 
 
-require 'validate/validate'
+require 'validator'
 
 
 task :validate_dash do
-  Fidor::SchemaValidation.main :dash
+  Validator.main :dash
 end
 
 task :validate_underscore do
-  Fidor::SchemaValidation.main :underscore
+  Validator.main :underscore
 end
 
 task :validate_jschema do
-  Fidor::SchemaValidation.main :jschema
+  Validator.main :jschema
 end
 
 task :validate_schema => [:validate_dash, :validate_underscore, :validate_jschema] do
-  
+
 end
