+++ lib/validator.rb
@@ -1,0 +1,4 @@
+require 'validator/base'
+require 'validator/json-schema'
+require 'validator/json_schema'
+require 'validator/j_schema'
