+++ spec/validator/json_schema_spec.rb
@@ -1,0 +1,8 @@
+require 'spec_helper'
+require 'validator'
+
+describe Validator::Json_Schema do
+  it 'runs' do
+    expect(true).to eq true
+  end
+end
